Check if the json file exists and create it if not

+++ main.rb
@@ -12,10 +12,16 @@ data = {
   "local": "earth"
 }
 
-serialized_data = JSON.pretty_generate(data)
 
-File.open(file, "w") do |file|
-  file.write(serialized_data)
+if File.exist?(file)
+  data = File.read(file)
+  deserialized_data = JSON.parse(data)
+  puts deserialized_data
+else
+  serialized_data = JSON.pretty_generate(data)
+  File.open(file, "w") do |file|
+    file.write(serialized_data)
+  end
 end
 
 def hello
